Build: 27x0: Fixed build issues for Aarch64 builds (#76)
AST2700 uses aarch64 compilers instead of arm32. Fixed some of the
compilation issues seen with early aarch64 builds where size_t is mapped
to 64bit variable instead of 32 bit variable.
Testing: Compiled successfully. Flashing BMC on AST2700 and dumping
version:

```
# REPOSITORY: pldmd +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# HEAD: d269047d78b093ad3026ec7b2106d3a44a17cfcb
# TAG: fatal: No names found, cannot describe anything.
# STATUS:
# ## devtool
#  M src/firmware_update.cpp
```

+++ src/firmware_update.cpp
@@ -565,7 +565,7 @@ int FWUpdate::processSendMetaData(const boost::asio::yield_context yield)
 
     // Calculate based on size of payload and maximum transfer size
     // Max number of requests including the requeries
-    size_t maxNumReq = findMaxNumReq(fwDeviceMetaData.size());
+    size_t maxNumReq = findMaxNumReq(static_cast<uint32_t>(fwDeviceMetaData.size()));
     // Max number of unique requests (excluding requeries)
     size_t numExpectedRequests = calcMaxNumReq(fwDeviceMetaData.size());
 
@@ -1447,7 +1447,7 @@ int FWUpdate::processSendPackageData(const boost::asio::yield_context yield)
     size_t offset = 0;
     int retVal = 0;
     size_t length = PLDM_FWU_BASELINE_TRANSFER_SIZE; // max payload size
-    const size_t dataSize = packageData.size();
+    const uint32_t dataSize = static_cast<uint32_t>(packageData.size());
 
     // Calculate based on size of payload and maximum transfer size
     // Max number of requests including the requeries
